Implmented sliding window solution
There are two main paths to this,
first we find the first and last overlapping window.
merge these with the new interval and insert this into the middle of
result skipping over all overlapping windows

if there are no overlapping windows we just need to insert
the interval into the correct position.